+++ shared/src/commonMain/kotlin/presentation/ui/main/home/HomeScreen.kt
@@ -82,13 +82,13 @@ import shoping_by_kmp.shared.generated.resources.special_for_you
 @OptIn(ExperimentalFoundationApi::class, ExperimentalResourceApi::class)
 @Composable
 fun HomeScreen(
-    navigateToDetail: (Int) -> Unit = {},
+    navigateToDetail: (Long) -> Unit = {},
     state: HomeState,
     events: (HomeEvent) -> Unit = {},
     navigateToNotifications: () -> Unit = {},
     navigateToSetting: () -> Unit = {},
     navigateToCategories: () -> Unit = {},
-    navigateToSearch: (Int?, Int?) -> Unit = { _, _ -> },
+    navigateToSearch: (Long?, Int?) -> Unit = { _, _ -> },
 ) {
